+++ Byndyusoft.ApiClient.sln
@@ -7,19 +7,19 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Byndyusoft.ApiClient", "src
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Byndyusoft.ApiClient.Tests", "tests\Byndyusoft.ApiClient.Tests.csproj", "{A4D9844E-C2DA-45EB-8AA8-1378B5072137}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Byndyusoft.ApiClient.Example.Server", "example\Api\Byndyusoft.ApiClient.Example.Server.csproj", "{D3BAEB82-7CEE-41C6-9368-94410FB532D8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "example\Api\Api.csproj", "{D3BAEB82-7CEE-41C6-9368-94410FB532D8}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "example", "example", "{8C97A966-5D6A-4D7F-9AC8-2D582883152A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Byndyusoft.ApiClient.Example.Contracts", "example\Api.Contracts\Byndyusoft.ApiClient.Example.Contracts.csproj", "{74DB2F65-EF05-444C-BA5A-38355B109393}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Contracts", "example\Api.Contracts\Api.Contracts.csproj", "{74DB2F65-EF05-444C-BA5A-38355B109393}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Byndyusoft.ApiClient.Example.Models", "example\Models\Byndyusoft.ApiClient.Example.Models.csproj", "{5E8DB677-D0A1-43FA-89A9-8CE7B1D08B1A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Models", "example\Models\Models.csproj", "{5E8DB677-D0A1-43FA-89A9-8CE7B1D08B1A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Byndyusoft.ApiClient.Example.Client", "example\Api.Client\Byndyusoft.ApiClient.Example.Client.csproj", "{CF40D501-21D9-4216-9129-E0A06A9D41B5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Client", "example\Api.Client\Api.Client.csproj", "{CF40D501-21D9-4216-9129-E0A06A9D41B5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Byndyusoft.ApiClient.Example.Tests", "example\Tests\Byndyusoft.ApiClient.Example.Tests.csproj", "{8321C290-1333-4FB1-BB85-C9958B1C98F1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "example\Tests\Tests.csproj", "{8321C290-1333-4FB1-BB85-C9958B1C98F1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Byndyusoft.ApiClient.Example.Benchmark", "example\Benchmark\Byndyusoft.ApiClient.Example.Benchmark.csproj", "{DAA94A70-F5EE-4912-B04E-4FF9FF74CDCE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Benchmark", "example\Benchmark\Benchmark.csproj", "{DAA94A70-F5EE-4912-B04E-4FF9FF74CDCE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
