+++ example/Api.Client/Api.Client.csproj
@@ -4,7 +4,7 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>Byndyusoft.ApiClient.Example.Client</RootNamespace>
+        <RootNamespace>Api.Client</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
@@ -15,6 +15,6 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\src\Byndyusoft.ApiClient\Byndyusoft.ApiClient.csproj" />
-      <ProjectReference Include="..\Api.Contracts\Byndyusoft.ApiClient.Example.Contracts.csproj" />
+      <ProjectReference Include="..\Api.Contracts\Api.Contracts.csproj" />
     </ItemGroup>
 </Project>
