+++ example/Api.Contracts/Api.Contracts.csproj
@@ -4,11 +4,11 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>Byndyusoft.ApiClient.Example.Models</RootNamespace>
+        <RootNamespace>Api.Contracts</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="protobuf-net.Core" Version="3.2.46" />
+      <ProjectReference Include="..\Models\Models.csproj" />
     </ItemGroup>
 
 </Project>
