+++ example/Benchmark/Benchmark.csproj
@@ -2,7 +2,7 @@
     <PropertyGroup>
         <TargetFramework>net8.0</TargetFramework>
         <OutputType>Exe</OutputType>
-        <RootNamespace>Byndyusoft.ApiClient.Example.Benchmark</RootNamespace>
+        <RootNamespace>Benchmark</RootNamespace>
     </PropertyGroup>
     <PropertyGroup>
         <PlatformTarget>AnyCPU</PlatformTarget>
@@ -18,7 +18,7 @@
         <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="8.0.11" />
     </ItemGroup>
     <ItemGroup>
-      <ProjectReference Include="..\Api.Client\Byndyusoft.ApiClient.Example.Client.csproj" />
-      <ProjectReference Include="..\Api\Byndyusoft.ApiClient.Example.Server.csproj" />
+      <ProjectReference Include="..\Api.Client\Api.Client.csproj" />
+      <ProjectReference Include="..\Api\Api.csproj" />
     </ItemGroup>
 </Project>
