+++ example/Benchmark/MvcTestFixture.cs
@@ -1,9 +1,9 @@
 #nullable enable
-namespace Byndyusoft.ApiClient.Example.Benchmark;
+namespace Benchmark;
 
 using System;
 using System.Net.Http;
-using Server;
+using Api;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.Hosting;
