+++ example/Models/PersonData.cs
@@ -1,4 +1,4 @@
-namespace Byndyusoft.ApiClient.Example.Models;
+namespace Models;
 
 using System.Collections.Concurrent;
 using System.Collections.Generic;
